PY20240718/tibame-group/requirements.txt
fitz==0.0.1.dev2
Flask==3.0.3
langchain_openai==0.1.17
openai==1.35.14
python-dotenv==1.0.1
